+++ src/main/java/com/example/demo/dal/IEntrepriseRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "entreprises-list", collectionResourceRel = "entreprises_data")
-public interface IEntrepriseRepository extends CrudRepository<Entreprise, Integer> {
+public interface IEntrepriseRepository extends CrudRepository<Entreprise, Long> {
 }
